chore: Add test.py to .gitignore
test.py is a file used for testing very small sections of code, new
ideas, functions and packages. The content of the file will always be
disposable due to the nature of its use. Anything that needs to be
retained will be added to a separate file or will be added as commented
code in the main run.py file.